Change README.md
Changed the text on some buttons to emoticons
Changed the localization file to json format
Changed the keys in the localization files to short words/phrases in English
Changed the functions to support the localization file and the required words in them
Added several other languages ​​to the localizations folder

Optimized the code and started making documentation for it.
Changed the embeds that are displayed when interacting with the bot.
Created new folders to implement new upcoming dating features.